Move Line class to a separate file
